Move code around, create executable for parsing files and create gemspec
