+++ .gitignore
@@ -1,0 +1,7 @@
+*.gem
+*~
+*#
+*#*
+.DS_Store
+.DS_Store?
+.irbrc
