+++ bin/mailbox
@@ -1,0 +1,19 @@
+#!/usr/bin/env ruby
+
+base = File.join(File.dirname(__FILE__), "..")
+$:.unshift File.join(base, "lib")
+
+require 'mailbox/interpreter'
+
+if ARGV.empty? || ARGV.first == "-h" || ARGV.first == "--help"
+  puts "Usage: mailbox prog_file.mail"
+  exit 0
+end
+
+filepath = File.expand_path(ARGV.shift)
+
+File.open(filepath) do |file|
+  Mailbox::Interpreter.parse_program(file.read)
+end
+
+Mailbox::Interpreter.run!
