+++ lib/mailbox/interpreter/version.rb
@@ -1,0 +1,5 @@
+module Mailbox
+  module Interpreter
+    VERSION = "0.0.1"
+  end
+end
